Add Google user authentication
Google user authentication is implemented using token validation,
so we don't need to actually ask the server for any user information.

We currently issue bearer tokens with the user's ID in them, and
validate them with a signature. If a user loses their token, they
will be compromised forever under this scheme.